scripts/requirements.txt
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
flatbuffers==25.12.19
idna==3.11
ImageIO==2.37.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
onnxruntime==1.24.4
opencv-python-headless==4.13.0.92
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
protobuf==7.34.1
PyMatting==1.1.15
referencing==0.37.0
rembg==2.0.69
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
sympy==1.14.0
tifffile==2026.4.11
tqdm==4.67.3
urllib3==2.6.3
